feat - 수업 추가 시 화면 이동 수정

+++ app/src/main/java/com/dogandpigs/fitnote/presentation/lesson/addlesson/AddLessonNavigation.kt
@@ -11,7 +11,7 @@ fun NavGraphBuilder.addAddLessonScreen(
     route: String,
     popBackStack: () -> Unit,
     navigateToLoadLesson: () -> Unit,
-    navigateToMemberLessonList: () -> Unit,
+    navigateToMemberLessonList: (Int) -> Unit,
 ) {
     val argument = ARGUMENT_MEMBER_ID
